添加发布产品产生保全链功能

+++ src/main/java/com/baoquan/shuxin/service/impl/user/UserServiceImpl.java
@@ -60,7 +60,7 @@ public class UserServiceImpl<T> implements UserService {
 
 	// 根据id获取用户的信息
 	@Override
-	public User findByIdUserInfo(Long id) {
+	public User findByIdUserInfo(Integer id) {
 		if (id == null) {
 			return null;
 		}
